fix: Refactor folder structure
Merge report generation folder structure to api-app
