+++ api-app/.env
@@ -1,1 +1,2 @@
 HYDROSENS_URL=http://localhost:5000/hydrosens
+GEMINI_API_TOKEN=AIzaSyAEs68XQOaFAk_qqgzm8rc0wi0JsfNCY-s
