+++ api-app/app.py
@@ -1,4 +1,5 @@
-from flask import Flask, request, jsonify
+from flask import Flask, request, jsonify, send_file
+from  utils.generate_report import run_generate_report
 import requests
 import os
 
@@ -45,3 +46,23 @@ def analyze():
     else:
         return jsonify({"error": "Invalid statistics parameter"}), 400
 
+
+@app.route('/generate-report', methods=['POST'])
+def generate_report():
+    try:
+        # Run report generation and get the output PDF path
+        pdf_file_path = run_generate_report(request.get_json())
+
+        # Send the file to the user
+        return send_file(
+            pdf_file_path,
+            as_attachment=True,
+            mimetype='application/pdf',
+            download_name='report.pdf'
+        )
+    except Exception as e:
+        app.logger.error(f"Report generation failed: {str(e)}")
+        return jsonify({"error": str(e)}), 500
+
+if __name__ == "__main__":
+    app.run(host="0.0.0.0", port=5000)
