+++ api-app/data/generated_reports/RSS_Hydro_Region_Report_2025.tex
