+++ api-app/requirements.txt
@@ -1,2 +1,6 @@
 flask
 requests
+Jinja2
+google-genai
+python-dotenv
+google-api-core
